Lab4-ReflectionAgent/CoverLetterBuilder/requirements.txt
streamlit
langgraph
langchain
langchain-core
langchain-community
pdfplumber
beautifulsoup4
requests
plotly
snowflake-connector-python
snowflake-snowpark-python
python-dotenv
fpdf2
python-docx